feat(setup): feature engineer, founder, and software-factory; fold the rest under More profiles
The first setup screen now shows only the featured default-catalog profiles
(engineer, founder, software-factory, in that order, when the catalog has
them), a "More profiles (N)" row that reopens the picker with every
non-abstract profile, and the import row. The full list is shown directly
when nothing is featured or the current default is not featured. Abstract
profiles are still excluded at discovery.

The setup extension tests are split into a shared harness, the walkthrough
suite, and the provider-step suite to stay under the file-length limit.
