+++ code/cli/tests/unit/setup.test.ts
@@ -71,11 +71,20 @@ afterEach(() => {
 });
 
 describe('setup state machine', () => {
-  it('discovers the default Outfitter catalog with display metadata, engineer first, abstract profiles hidden', () => {
+  it('discovers the default Outfitter catalog with featured profiles first in order, abstract profiles hidden', () => {
     const { catalog } = createTree();
+    write(join(catalog, 'agents', 'zeta', 'agent.md'), '---\nname: zeta\ndescription: Last.\n---\n');
+    write(join(catalog, 'agents', 'planner', 'agent.md'), '---\nname: planner\ndescription: Plans.\n---\n');
+    write(
+      join(catalog, 'agents', 'software-factory', 'agent.md'),
+      '---\nname: software-factory\ndescription: Ships.\n---\n',
+    );
     expect(discoverSetupAgentChoices({ defaultCatalogRoot: catalog })).toEqual([
-      { id: 'engineer', label: 'Engineer', description: 'Engineering profile.' },
-      { id: 'founder', label: 'Founder', description: 'Founder/operator profile.' },
+      { id: 'engineer', label: 'Engineer', description: 'Engineering profile.', featured: true },
+      { id: 'founder', label: 'Founder', description: 'Founder/operator profile.', featured: true },
+      { id: 'software-factory', label: 'software-factory', description: 'Ships.', featured: true },
+      { id: 'planner', label: 'planner', description: 'Plans.' },
+      { id: 'zeta', label: 'zeta', description: 'Last.' },
     ]);
     expect(discoverSetupAgentChoices({})).toEqual([]);
   });
