+++ docs/architecture/README.md
@@ -259,7 +259,7 @@ Outfitter launches `codex`, maps model selection to `-m`, and injects selected M
 - **`outfitter run [agent]`** (default command): resolve → compose → project → launch.
   A positional agent slug selects what to run (default from settings `default_agent`); `--harness <pi|claude|codex>` selects the harness (default from settings `default_harness`, then `pi`); unknown args pass through to the child CLI; `--strict` makes warnings fatal.
   Interactive clean-home launches start Pi-native onboarding; non-interactive clean-home launches require `outfitter setup` first.
-- **`outfitter setup [source]`**: launch the bundled, offline Pi walkthrough: one profile screen (default-catalog profiles plus an import-another-catalog row), original target wording, and optional direct-source path; append only the default CLI-agent choice (Pi/Outfitter preselected) and, for Pi with no provider connected, an in-session `/login` step; then atomically apply the result through the CLI state machine.
+- **`outfitter setup [source]`**: launch the bundled, offline Pi walkthrough: one profile screen (featured default-catalog profiles, a More-profiles row, and an import-another-catalog row), original target wording, and optional direct-source path; append only the default CLI-agent choice (Pi/Outfitter preselected) and, for Pi with no provider connected, an in-session `/login` step; then atomically apply the result through the CLI state machine.
 - **`outfitter sync`**: validate local settings; atomically fetch configured remote settings; reload the merged settings; then atomically fetch every resulting remote source into `<cache_directory>/repos/<encoded-uri-and-ref>/` and record its verified commit under `<cache_directory>/source-state/`.
   Validate before swapping, preserve the last good checkout on failure, report per-source status, and redact embedded credentials from all output. Required-source failures exit nonzero.
 - **`outfitter run`**: enforce the selected source-cache policy before resolution. Healthy manifest-backed caches stay network-free; repair mode atomically rebuilds unhealthy remote caches, locked mode requires full commit pins, and offline mode fails closed without network access. Explicit local paths are never cache-managed.
